+++ src/components/Kpi/Kpies.tsx
@@ -1,76 +1,11 @@
 import { useMemo } from "react";
-import { assignedStatusEnum, patientsStatus } from "../functions/patientsStatus";
-import useGetShuttleDetailsPerRequest from "../hooks/data/useGetShuttleDetailsPerRequest";
-import useGetShuttleRequests from "../hooks/data/useGetShuttleRequests";
-import useGetShuttles from "../hooks/data/useGetShuttles";
-import Kpi from "./Kpi";
-import SummarizeNumbers from "./SummarizeNumbers";
+import { assignedStatusEnum, patientsStatus } from "../../functions/patientsStatus";
+import useGetShuttleDetailsPerRequest from "../../hooks/data/useGetShuttleDetailsPerRequest";
+import useGetShuttleRequests from "../../hooks/data/useGetShuttleRequests";
+import useGetShuttles from "../../hooks/data/useGetShuttles";
+import SummarizeNumbers from "../SummarizeNumbers";
+import getMatchingKpi from "./KpiGet";
 
-interface IGetMatchingKpi {
-  title: string,
-  value: string | number | undefined,
-  inline: boolean,
-}
-export const getMatchingKpi = ({ title, value, inline }: IGetMatchingKpi) => {
-  switch (title) {
-    case "טרם שובצו":
-    case "טרם שובץ":
-    case "ממתינים לחזור":
-    case "ממתין לחזור":
-      return <Kpi
-        title={title}
-        value={value}
-        titleColor="--Colors-orange-6"
-        borderColor="--Colors-orange-3"
-        inline={inline}
-      />
-
-    case "שובצו":
-    case "שובץ":
-      return <Kpi
-        title={title}
-        value={value}
-        titleColor="--Colors-purple-6"
-        borderColor="--Colors-purple-3"
-        inline={inline}
-      />
-
-    case "נשלחה הודעה":
-    case "במרפאה":
-    case "חזרו":
-    case "חזר":
-      return <Kpi
-        title={title}
-        value={value}
-        titleColor="--Colors-green-6"
-        borderColor="--Colors-green-3"
-        inline={inline}
-      />
-
-    case "מגיעים בקרוב":
-    case "מגיע בקרוב":
-      return <Kpi
-        title={title}
-        value={value}
-        titleColor="--Colors-cyan-6"
-        borderColor="--Colors-cyan-3"
-        inline={inline}
-      />
-
-    case "מסיימים בקרוב":
-    case "מסיים בקרוב":
-      return <Kpi
-        title={title}
-        value={value}
-        titleColor="--Colors-geekblue-6"
-        borderColor="--Colors-geekblue-3"
-        inline={inline}
-      />
-
-    default:
-      return ""
-  }
-}
 
 const Kpies = () => {
   const shuttleRequests = useGetShuttleRequests();
